pony-data-collector/requirements.txt
yt-dlp>=2024.1.0
opencv-python>=4.9.0
pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
supabase>=2.0.0
requests>=2.31.0
